requirements.txt.txt
yfinance
pandas
numpy
matplotlib
pytrends
requests
pyyaml
joblib
streamlit
